--- src/livestream/index.js
@@ -1,1 +1,0 @@
-export MediaPlayer from './MediaPlayer';
