+++ src/media-player/controls/MediaPlayerSafeLayout.js
