+++ src/media-player/controls/Seeker/Timestamp.js
