+++ src/media-player/controls/Seeker/index.js
